fixed a bug where the settings view controller would incorrectly obscure a UILabel.
